+++ baybe/surrogates/gaussian_process/presets/__init__.py
@@ -1,9 +1,9 @@
 """Gaussian process surrogate presets."""
 
-# Criterion
+# Core
 from baybe.surrogates.gaussian_process.components.fit_criterion import FitCriterion
 
-# Default preset
+# BayBE preset
 from baybe.surrogates.gaussian_process.presets.baybe import (
     BayBEFitCriterionFactory,
     BayBEKernelFactory,
@@ -23,8 +23,6 @@ from baybe.surrogates.gaussian_process.presets.chen import (
     CHEN_FIT_CRITERION_FACTORY,
     ChenKernelFactory,
 )
-
-# Core
 from baybe.surrogates.gaussian_process.presets.core import GaussianProcessPreset
 
 # EDBO preset
@@ -35,7 +33,7 @@ from baybe.surrogates.gaussian_process.presets.edbo import (
     EDBOMeanFactory,
 )
 
-# Smoothed EDBO preset
+# EDBO Smoothed preset
 from baybe.surrogates.gaussian_process.presets.edbo_smoothed import (
     SMOOTHED_EDBO_FIT_CRITERION_FACTORY,
     SmoothedEDBOKernelFactory,
@@ -43,11 +41,18 @@ from baybe.surrogates.gaussian_process.presets.edbo_smoothed import (
     SmoothedEDBOMeanFactory,
 )
 
+# Hvarfner preset
+from baybe.surrogates.gaussian_process.presets.hvarfner import (
+    HvarfnerKernelFactory,
+    HvarfnerLikelihoodFactory,
+    HvarfnerMeanFactory,
+)
+
 __all__ = [
     # Core
     "FitCriterion",
     "GaussianProcessPreset",
-    # Default BayBE preset
+    # BayBE preset
     "BayBEFitCriterionFactory",
     "BayBEKernelFactory",
     "BayBELikelihoodFactory",
@@ -64,9 +69,13 @@ __all__ = [
     "EDBOKernelFactory",
     "EDBOLikelihoodFactory",
     "EDBOMeanFactory",
-    # Smoothed EDBO preset
+    # EDBO Smoothed preset
     "SMOOTHED_EDBO_FIT_CRITERION_FACTORY",
     "SmoothedEDBOKernelFactory",
     "SmoothedEDBOLikelihoodFactory",
     "SmoothedEDBOMeanFactory",
+    # Hvarfner preset
+    "HvarfnerKernelFactory",
+    "HvarfnerLikelihoodFactory",
+    "HvarfnerMeanFactory",
 ]